Fix off-by-one in _renderTimeRange
Previously _renderTimeRange was always used for ActiveHours, where the active hours (1, 2) would imply the second hour fully meets the condition so the end of the time range would be hour 3.
This is not the case for continuous insights, which end at the end of the time range.
Make _renderTimeRange stop applying +1 offsets, and apply them only in the ActiveHours case